more type fixes

+++ misc.c
@@ -2548,7 +2548,7 @@ GC_API void GC_CALL GC_set_time_limit(unsigned long value)
 
 GC_API unsigned long GC_CALL GC_get_time_limit(void)
 {
-    return GC_time_limit / 1000000;
+    return (unsigned long)(GC_time_limit / 1000000);
 }
 
 GC_API void GC_CALL GC_set_time_limit_ns(unsigned long long value)
